+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Treat `jcs-eddsa-2022` and `eddsa-jcs-2022` cryptosuites as different.
 - Use `jcs-eddsa-2022` cryptosuite for signing activities.
 - Improved object type disambiguation in `Update()` activity handler.
+- Process all objects with `attributedTo` property as notes.
 
 ### Fixed